ITC18: Fix erroneous string comparison.
It was probably working, but if so only incidentally because of compiler
specifics (using a single instance of equal constant strings) and
std::string implementation details (returning the c_str() from which it
was constructed).

I suspect the real intension was to use g_protocol_file_available, which
is defined but never used. I'm leaving this unfixed for now.