Prior: Add speed control for Z (Korbinian Kapsner).
Date: Mon, 23 Sep 2013 16:02:47 +0200
From: Korbinian Kapsner <korbinian.kapsner@ph.tum.de>
To: micro-manager-general@lists.sourceforge.net
Subject: [micro-manager-general] Added functionality to Prior z-stage
	controller

Hi all,

the Prior z-stage controller has some serial commands to specify the
maximum speed, acceleration and s-curve. These are not implemented in
the .dll for micro-manager unlike the commands for the xy-stage.

Since the commands are quite the same like for the xy-stage I copied the
code for the xy-stage to the z-stage (just changing the command names:
SAS -> SAZ, SCS -> SCZ, SMS -> SMZ) and it seems to work.

I attach the modified code.

Hope this will help somebody sometime.

best
Korbinian

PS: By looking at the specs it seemed to me that the maximal values for
SAS, SCS and SMS (and now also the ...Z) are not correct. As far as I
understand it they all lie between 1 and 100. But I only checked for the
ProScan III... If anybody has a deeper insight please correct the upper
limit.