MMCorePy: Remove nonfunctional code.
The intention of the code removed here was presumably to override the 'message'
attribute of the Python wrappers of the exception objects. However, (A) it does
not work, because BaseException's message attribute cannot be overridden by
this mechanism (so these functions were never called, at least in recent
versions of Python), and (B) it was resulting in invalid code generated by SWIG
(taking the address of a temporary object), resulting in compiler errors with
recent versions of GCC (and warnings elsewhere).

Issue (B) can be fixed by replacing std::string with const char *const, but I'm
just removing it all because it does not actually do anything, anyway.

By the way, BaseException.message was deprecated in Python 2.6 and is obsolete.