SpectralLMM5: Remove what is almost certainly broken code.
There was this 'readWriteSame_' mechanism (see micromanager1.3 r4166), which
looks like a hack to work around a crash. Based on the documentation, there is
no reason why it should work. It is possible that this workaround was devised
based on experimentation, because, at least in the case of the 0x23 and 0x26
commands, it so happens that a '0x23' without the required parameters can
return whatever bytes that were set by previous commands (even if those
commands are not 0x23 commands).

Since there is no information left about why this was done, and it is clearly
not doing the right thing with at least two LMM5 controllers tested, I am
removing the mechanism and replacing it with a simple check for a correct
response to a 0x26 command. The tested hardware, both of which do not appear to
support trigger-out, reply to the 0x26 command with an empty string. In these
cases, we do not make the trigger-related properties available. I cannot
presently test the case where triggers _are_ supported by the device; that will
have to be fixed (if necessary) when somebody has a device available for
testing (or reports that their system brakes with this change). Incidentally,
Spectral is working on improving the trigger capabilities, so the command set
may change in the near future.