dc1394: Fix invalid property value handling.
Unlikely to be enoucntered, but error handling could have been being
skipped for an invalid value for framerate or video mode. (The
dc1394framerate_t and dc1394video_mode_t are enum types which can be
unsigned, so using -1 as an indicator of invalid value does not work.)

Pointed out by clang++ warnings.