TIFF Writer: Use correct byte count, not character count.
The bug was causing corrupted TIFF files to be written when the locale was set
to e.g. zh_CN, because the matadata (and potentially other strings) could
contain non-ASCII characters whose UTF-8 representation is more than 1
byte/character. It could cause the string to overwrite neighboring data
structures (which happens to be the first IFD in the case of the summary
metadata string!).