New, cleaner search path mechanism for device adapters.
Add setDeviceAdapterSearchPaths(), getDeviceAdapterSearchPaths(), and
getDeviceAdapterNames() to the Core. These allow client code to completely
determine the paths in which device adapters are searched for, while providing
defaults that make sense.

Code that uses addSearchPath() and getDeviceLibraries() will still work, as
long as the same code does not call the new functions.