Allow better control over where to load MMCoreJ_wrap from.
If the Java system property mmcorej.library.path is set, use that path and only
that path. Otherwise, fall back to a backward-compatible mechanism, but with
better logging.

Also, do not indiscriminately add a bunch of paths to the device adapter search
path; allow the Core to determine the default search paths.