SerialManager: Avoid breaking up writes into bytes.
When writing a sequence of characters, hand the whole buffer to
Boost.Asio (and hence to the operating system), instead of sending
length-1 buffers in a loop. This only applies to the case where
DelayBetweenChars is set to zero (< 0.001).

The reasons for this change are that (1) it does not appear that there
was any reason to split the messages into characters and (2) some USB
devices (that present a serial interface via their driver) may not
function correctly if a command is sent as multiple USB packets, and
sending off whole command messages to the OS is likely to avoid such
issues.

See:
http://micro-manager.3463995.n2.nabble.com/Neos-configuration-td7582451.html