Fix hacks involving CDeviceBase [8/8]: ASITiger.
This final case is more complicated than the 7 others in this series of
8 patches.

The gist of this refactoring is that the combination of ASIDevice and
ASIDeviceBase is replaced by two class templates, ASIBase and
ASIPeripheralBase, each of which implements common functionality.

ASIHub inherits from ASIBase<HubBase, ASIHub>, while all peripheral
devices inherit from ASIPeripheralBase<T, U> (which in turn inherits
from ASIBase<T, U>).

Since ASIBase and ASIPeripheralBase are class templates, all of their
methods are defined inline. However, they are straightforwardly
transplanted from the former ASIDevice.cpp.

The former ASIDevice::Initialize(bool) was renamed to
ASIPeripheralBase<>::PeripheralInitialize(bool) (since it does nothing
in the case of ASIHub), so as not to shadow the MM::Device method
Initialize(), and to clarify its role.