Minor fix for Clojure class loading
Previously, the context class loader was being set (1) for the EDT (the
thread on which the MMStudioMainFrame constructor is called) by the
method loading the acquisition engine (the mysteriously named
"startLoadingPipelineClass()") and (2) for the plugin loading thread by
the plugin loading thread.

With this change, the context class loader is set for the loading
threads (and only the loading threads). This clarifies the requirement
for Clojure class loading, and also avoids setting the EDT's context
class loader, which is global state.