Fix memory leak when running acquisitions.
VirtualAcquisitionDisplay now splits the "name" into "acquisition name" and
"window title" fields which are tracked separately. In turn this means that
when the VAD is closed, it has the acquisition name handy to close the
associated acquisition, even when that acquisition had an autogenerated (thus,
not desirable to be displayed to the user) name.