Ant: Prevent `make install' from touching Clojure JAR
In the no-AOT method (currently only used by SlideExplorer2), avoid
using a temporary directory and instead place intermediate files in a
defined location (scratchdir), so that rerunning the Ant compile and jar
targets does not change the timestamp of the JAR file when there are no
modifications. This is important, since `make install' implies `make
all', yet we do not want to touch an already-built tree when installing.