Don't pretend I know the optimal struct size
The start of vector may not be cache-aligned anyway.
Also decrease packet size to 127, since most lines will fit in it
(really this value should be determined by surveying actual logs; it is
somewhat important for keeping down memory usage when copious logging is
taking place on a slow output (Windows console)).