Fix erroneous canvas sizes at certain zoom levels.
When using a zoom level with a repeating decimal (e.g. 33.3333..%), the
ImageCanvas would allocate itself insufficient amounts of screen real estate
to display the entire image, causing the zoom indicator overlay to get drawn.
Our "fix" for this is to override DisplayWindow.pack() and force the canvas
to be larger prior to repacking.