MMCore logging: Avoid copy upon buffer swap
It looks like calling std::swap() on a pair of boost::container::vector
results in a deep copy of the data (log entry content), due to the
default std::swap() implementation being used. Use the member swap()
instead, which just swaps pointers, as we originally intended. In
particular, this avoids holding the lock (which synchronizes logging)
for an unnecessarily long time.