Pad script area line count by 1 to avoid scrolling the cursor offscreen.
Fixes ticket #721. Note that this does occasionally mean that the script area
will display a blank line with no line number (in the rare circumstance that
the old behavior that this replaces was actually the correct behavior). This
is just a quick patch job, and seems to be correct in something like 75% of
cases.