acqEngine: Fix bug causing spurious setting of Z position
In a time-channel acquisition where the channels are hardware-triggered
and the interval is large enough that separate time points are run as
separate bursts, the Z position was being set between each time point
for no good reason.

The bug was that the last-stage-positions field of the state map was
being set to null after each burst lacking a hardware-triggered Z stack.

Regression test added.