Fix bug causing spurious increase in size of the script area.
This does come at the cost of some additional computation -- now any time text
is added to the script, the entirety of the script contents are examined to
count the number of newlines. But I don't notice any slowdown on my computer,
even for gratuitously large (>1MB) scripts, so I doubt this is a big deal.