Rework the pack() override in DisplayWindow.
This fixes a bug causing the DisplayWindow to be gigantic when showing large
images (e.g.2kx2k), and also simplifies the logic somewhat. We do still have
a bug lurking around that causes the canvas to show its "zoom indicator" when
the zoom level is 33% or 16% because the canvas is 2px too small to fit.
However, the current behavior is a definite step up from the old behavior.