PICAM: Fix crash on setting certain properties
In practice, the crash was observed with the "Port" (readout port) property.
The cause was a confusion between PICAM enum values and the index of enum
values, leading to an out-of-bound array index.

There appears to be a separate issue that prevents correct setting of the
"Port" property, which is not addressed in this commit.