PICAM: Prevent failure to set EM gain
This prevents the EM gain property from reverting to the original value when
attempting to set it while an acquisition is _not_ running. It is a partial
fix, in that setting the property while running Live in MMStudio does not yet
work, and it has not been confirmed whether the set value is correctly
reflected in the next acquisition.

Change contributed by Toshio Suzuki.