[Display] Post DisplayCreatedEvent when display is created.
Requested by Nico for fixing an issue with the pattern overlay plugin.
This event is not exposed in the API, because it in turn exposes the
VirtualAcquisitionDisplay and DisplayWindow objects, which aren't in the API.
Functionally it's not very useful unless you also depend on those objects,
which means you're sidestepping the API to begin with.