Avoid discarding first image in Live acquisition
We were waiting for the first image before setting up the display, but
then discarding that first image. This could result in a black image
canvas being displayed for Live with very long exposures.

This is a quick fix, and there is more to be improved as the added
comments indicate.