[TIFF writing] Fix an assumption that was broken by a prior commit.
r14653 ensured that we never created a TaggedImage with null metadata, to fix
a downstream problem. However, the fixIndexMap() function explicitly checks for
null metadata to detect placeholder images. It has now been modified to also
check for zero-length metadata and also treat that as a placeholder; in
practice I do not think we will ever have real images with zero-length
metadata, so this should be safe, especially as fixIndexMap() is rarely
called.