[Display] Fix an error causing new images not to be shown.
Our scrollbars were telling the ImagePlus to change its position to one that
it thought was out-of-bounds (e.g. display frame 5 when the ImagePlus thinks
it has 4 frames). We now update its sizing information prior to doing this.
This fixes the bug where clicking the "-> Album" button would always show
the *previous* image in the album, instead of the one you just took.