IIDC: Add some color camera support
This should handle cameras that output RGB or YUV data directly. Cameras
that send raw images (to be debayered) are not yet supported.

Also add a workaround that allows capture to stop without error on OS X
(testing with the old FireWire Apple iSight camera). With this
particular camera, there is still a crash that happens almost always
when attempting to do a second capture after starting Micro-Manager. On
the surface, it looks like an OS X specific issue in libdc1394, which I
do not have the time to look into now. It is possible that different
call orders must be used for libdc1394 functions depending on the
platform, but I would rather test with a "real" camera (not the iSight)
if going that route.