[Multipage TIFF] Fix memory leak with large position lists.
Each individual MultipageTiffWriter was storing a full copy of the summary
metadata in string form; when a lot of such Writers are around (e.g. because
of a large position list) and the summary metadata is large (e.g. because of
a large position list), this can result in a lot of leaked memory. This
change centralizes storage of the summary metadata string into the
TaggedImageStorageMultipageTIFF object instead, so there's only one copy.