SplitView: fix running in MDA.
This modifies the acquisition logic to post an event that allows third-party
code to modify the summary metadata of acquisitions before it is used for
anything in the Java layer. Consequently the SplitView plugin can adjust the
number of channels and the size of the images, which fixes issues that prevented
SplitView acquisitions from displaying or saving properly.